added traits for each implemented interface
